+++ src/TokensTracker/TokensTracker/config.json
@@ -1,9 +1,7 @@
 {
 	"PluginConfiguration": {
-		"Nep17Path": "Nep17BalanceData",
-		"Nep11Path": "Nep11BalanceData",
+		"DBPath": "TokenBalanceData",
 		"TrackHistory": true,
-		"RecordNullAddressHistory": false,
 		"MaxResults": 1000,
 		"Network": 860833102
 	},
